microatf: ensure that struct atf_tc_s is large enough
When running the tests on a CheriBSD system on Arm Morello, pointers are
16 bytes, so the implicit size of 8192 bytes is not large enough to hold
all of struct atf_tc_impl_s_ and running the tests fails with a bounds
violation. To fix this issue, this commit explicitly sizes the array
instead of relying on the alignment and adds a static assertion to the
source file to ensure this invariant holds. I've also reordered
struct atf_tc_impl_s_ to avoid padding on CHERI systems.